scripts/do-bulk/requirements.txt
bitarray==1.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.1
idna==2.10
jsonpickle==2.0.0
parsimonious==0.8.1
pycryptodome==3.10.1
python-digitalocean==1.16.0
requests==2.25.1
rlp==2.0.1
six==1.16.0
toolz==0.11.1
urllib3==1.26.5
